backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pymongo==4.9.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
pydantic[email]==2.10.4
python-dotenv==1.0.1
openai==1.57.0
httpx==0.27.0
